+++ src/App.vue
@@ -1,14 +1,24 @@
 <!--
  * @Date: 2020-03-26 15:04:25
  * @Author: junfeng.liu
+<<<<<<< HEAD
  * @LastEditTime: 2020-04-30 16:44:18
+=======
+ * @LastEditTime: 2020-05-07 18:22:22
+>>>>>>> development/l_loading
  * @LastEditors: junfeng.liu
  * @Description: des
  -->
 <template>
     <div>
+<<<<<<< HEAD
         <div style="padding: 20px;width: 500px;background: #f6f6f6;position: relative;">
             <!-- <LLoading></LLoading> -->
+=======
+        <div style="padding: 20px;width: 500px;height: 300px;background: #f6f6f6;position: relative;">
+            <LMenu></LMenu>
+            <!-- <LLoading transfer></LLoading> -->
+>>>>>>> development/l_loading
             <!-- <LButtonGroup shape="circles" vertical size="small">
                 <LButton type="dashed" icon="logo-facebook" disabled></LButton>
                 <LButton type="primary" icon="ios-sunny-outline" disabled></LButton>
@@ -27,7 +37,7 @@
                 <Select slot="append" size="default"></Select>
             </LInput>
             <!-- <LButton type="info" shape="circles" icon="ios-crop" size="small" to="/hello" target="_blank" debounce earlyTrigger @click="handleClick">sadf</LButton> -->
-            <!-- <LButton type="cool-hover" shape="circle" icon="ios-crop" size="" debounce earlyTrigger @click="handleClick">sadf</LButton> -->
+            <LButton type="cool-hover" shape="circle" icon="ios-crop" size="" debounce earlyTrigger @click="handleClick">sadf</LButton>
         </div>
         <!-- <LLoading></LLoading> -->
         <!-- <LInput type="textarea" :autosize="{minRows: 3}" disabled></LInput>
@@ -86,7 +96,11 @@ export default {
             // this.show = true
             // this.range = ['asdf', 20]
             console.log('aaa')
-            this.disabled = !this.disabled
+            // this.disabled = !this.disabled
+            this.$loading.show()
+            setTimeout(() => {
+                this.$loading.hide()
+            }, 1000)
         }
     }
 }